Processing RIC Indication trigger

+++ e2sim/e2sm_examples/kpm_e2sm/src/ASN1c/RICrequestID.h
@@ -21,7 +21,7 @@ extern "C" {
 
 /* RICrequestID */
 typedef struct RICrequestID {
-	int	 ricRequestorID;
+	long	 ricRequestorID;
 	long	 ricInstanceID;
 	/*
 	 * This type is extensible,
